Update customizer
As it turns out, it's not just empty dictionaries on the right-hand side that get their references broken, it's _all_ dictionaries.  I got convinced otherwise by doing some bad tests in the console.  This new code is a bit simpler: if the left-hand value is not a POJO, or it's an empty POJO, replace it with a reference to the right-hand value.  For our use case, this is fine -- we don't care about weird instantiated objects in the left-hand side of config file merges.